Separate Model From Solver
The big performance problem at present is not being able to reuse the
model, as building it in the first place is quite expensive with all the
mallocs happening.  We separate that out, allowing it to be reset in a
quick way, and change how unit clauses are handled e.g. just set the
variable.  Also we remove a bunch of templating and use interfaces when
we want to use generic algorithms with type specific models.