Fix Unary Clauses (#10)
Before this it was possible to set v0 to true, and have a clause ~v0.
The way it workd was it would see the clause as unary, because no
notification happened, and it would set it to false and be satisfiable
erroneously.  As unary clauses are considered initial-state, we should
add them to the learned clauses so they are not persisted across model
reuse.